change int to double
main.cpp added

+++ calculator.h
@@ -4,8 +4,8 @@
 class Calculator
 {
 	public:
-		int Add (double, double);
-		int Sub (double, double);
+		double Add (double, double);
+		double Sub (double, double);
 };
 
 #endif//CALCULATOR_H
